🟡(change): refactor autocomplete field
broke the autocomplete field out into multiple components, we still need to implement caching on this to ensure the data is not consistently being fetched when it does not need to.
